Enhance category and product management features
- Updated permissions in `settings.local.json` for Bash commands.
- Modified project structure in `ArtStore.sln` and updated GUIDs.
- Refined `IExcelService.cs` and `Result.cs` for better handling.
- Changed cache keys in `ProductCacheKey.cs` for pagination and search.
- Added `IsActive` and `CategoryId` properties in `AddEditProductCommand.cs`.
- Implemented validation for category selection in `AddEditProductCommandValidator.cs`.
- Enhanced data retrieval in `GetAllProductsQueryHandler.cs`.
- Transitioned entity classes to use generics for IDs.
- Introduced new classes for category management with corresponding commands and validators.
- Updated UI components to support new features for categories and products.
- Modified API controllers to handle new commands and queries effectively.
- Updated project references and dependencies in `ArtStore.UI.Client.csproj`.
- Improved user experience in Razor components with better validation and loading indicators.

+++ src/ArtStore.Shared/DTOs/Order/Commands/UpdateOrderStatusCommand.cs
@@ -5,7 +5,7 @@ namespace ArtStore.Shared.DTOs.Order.Commands;
 
 public class UpdateOrderStatusCommand : ICommand<Result>
 {
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
     public OrderStatus Status { get; set; }
     public string? Notes { get; set; }
 }
