+++ src/ArtStore.Shared/DTOs/Order/OrderDetailDto.cs
@@ -2,8 +2,8 @@
 
 public class OrderDetailDto
 {
-    public int Id { get; set; }
-    public int OrderId { get; set; }
+    public long Id { get; set; }
+    public long OrderId { get; set; }
     public int ProductId { get; set; }
     public string? ProductName { get; set; }
     public int Quantity { get; set; }
