+++ src/ArtStore.Shared/DTOs/Order/OrderStatusHistoryDto.cs
@@ -4,8 +4,8 @@ namespace ArtStore.Shared.DTOs.Order;
 
 public class OrderStatusHistoryDto
 {
-    public int Id { get; set; }
-    public int OrderId { get; set; }
+    public long Id { get; set; }
+    public long OrderId { get; set; }
     public OrderStatus? OldStatus { get; set; }
     public OrderStatus NewStatus { get; set; }
     public DateTime ChangedAt { get; set; }
